alt_git/requirements.txt
fastapi
pydantic
deepdiff
GitPython
httpx
